+++ src/Eorm/Library/Storage.php
@@ -131,7 +131,7 @@ class Storage
                 $argument
             );
 
-            $this->primaryKeys = Helper::range((int) Server::insertId($this->server), $rows, true);
+            $this->primaryKeys = Helper::range(Server::insertId($this->server), $rows, true);
 
             return $this->reload();
         } else {
@@ -199,7 +199,7 @@ class Storage
         );
 
         $this->primaryKeys = Helper::merge(
-            Helper::range((int) Server::insertId($this->server), $rows, true),
+            Helper::range(Server::insertId($this->server), $rows, true),
             $this->primaryKeys
         );
